v-omni-ai/v-omni-embedding/requirements.txt
numpy
onnx
onnxruntime-gpu
pillow
transformers